[FIXED] NatsBench
The tool is now inline with the other NATS clients bench tools.

The number of messages given in the command line is the total
number of messages to be sent. So if there is more than one publisher,
the per-publisher number is the number of messages divided
by the number of publishers (`-np`). The last publisher may get a
slight different number of messages to send due to the possible
fraction.

Each subscription is supposed to receive each message (pub/sub),
therefore each subscription receives the total number of messages
as given in the command line (`-n`).

Resolves #133