Converted read for msg line to char array instead of char buffer.
Cleaned up some tests.

Fixed java doc in gradle and in public API for initial build to work. Needs more.

There is still an intermittent bug affecting request/reply "sometimes". It may be related
to this commit, but maybe not. it is very sneaky and requires lots of test runs to initiate.