Fixed a nasty bug in NatsConnectionReader
There was an issue where the reader didn't reset the opPos and
gotCR variables on disconnect. This wasn't tested, so I also added
a test to force a bad protocol line that is missing the \n after a \r.

Tested with an external tool as well.

Also, added comments to publish to point to the send buffer settings and
highlight the possible exception when that buffer is full.

Fixes #186
Fixes #191
Fixes nats-io/java-nats-streaming#107