Refactored MessagesQueue to use LinkedBlockingQueue
message queue was using a lot of custom code to handle all the performance and locking requirements. Now it uses a LinkedBlockingQueue which is much simpler. The new implementation uses a poison pill to shutdown.

By switching to a blocking queue we get back pressure on publishers. This changes the performance characteristics from Autobench but mostly to smooth them out over all message sizes. Latency and request/reply performance are nominal.

Added error handler to NatsBench.
Fixed a minor warning in raw tcp test

Prepare for 2.5.0, updated all versions