Allow configuring a message handler to be executed before a message is published
Timestamping messages in client code yields spurious results, because there is internal NATS
queueing involved. Consequently, any GC or safepoint pause which happens between the time
a message is equeued and then dequeued gets incorporated into the delay attributed to NATS.

This commit introduces a custom message handler to be executed before a message is published.
This allows reducing the time gap curretly assinged to the message broker.